Ready to publish alpha 1.0

+++ BlogSite/Controllers/UsersController.cs
@@ -22,7 +22,7 @@ namespace BlogSite.Controllers
         }
 
         // GET: Users/Details/5
-        public ActionResult Details(decimal id)
+        public ActionResult Details(int id)
         {
             if (id == null)
             {
